# docs/swagger.yaml
definitions:
  github_com_maczh_mgin_models.Result-any:
    properties:
      data: {}
      msg:
        type: string
      page:
        $ref: '#/definitions/models.ResultPage'
      status:
        type: integer
    type: object
...
  github_com_maczh_mgin_models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
...
  github_com_maczh_mgin_models_sys.SysUser:
    properties:
      avatar:
        type: string
      createAt:
        type: string
      createBy:
        type: string
      delFlag:
        type: integer
      email:
        type: string
      id:
        type: integer
      loginDate:
        type: string
      loginIp:
        type: string
      loginName:
        type: string
      mobile:
        type: string
      nickName:
        type: string
      password:
        type: string
      remark:
        type: string
      sex:
        type: integer
      status:
        type: integer
      tenantId:
        type: integer
      updateAt:
        type: string
      updateBy:
        type: string
      userType:
        type: string
    type: object
...
  github_com_maczh_mgin_models_sys_request.ChangePasswordReq:
    properties:
      loginName:
        type: string
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - loginName
    - newPassword
    - oldPassword
    type: object
  github_com_maczh_mgin_models_sys_request.ChangeStatusReq:
    properties:
      id:
        type: integer
      status:
        type: integer
    required:
    - id
    - status
    type: object
...
  github_com_maczh_mgin_models_sys_request.DeleteByIdReq:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  github_com_maczh_mgin_models_sys_request.LoginReq:
    properties:
      loginName:
        description: 可以是用户名或邮箱或手机号，自动识别
        type: string
      password:
        type: string
    required:
    - loginName
    - password
    type: object
  github_com_maczh_mgin_models_sys_request.RegisterReq:
    properties:
      avatar:
        type: string
      email:
        type: string
      loginName:
        type: string
      mobile:
        type: string
      nickName:
        type: string
      password:
        type: string
      sex:
        type: integer
    required:
    - loginName
    - password
    type: object
...
  github_com_maczh_mgin_models_sys_request.VerifyTokenReq:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  models.ResultPage:
    properties:
      count:
        description: 总页数
        type: integer
      index:
        description: 页号
        type: integer
      size:
        description: 分页大小
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 已注册用户进行登录操作
      parameters:
      - description: 用户登录请求参数
        in: body
        name: loginReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回 token
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或登录失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 用户登录
      tags:
      - 用户管理
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: 新用户进行注册操作
      parameters:
      - description: 用户注册请求参数
        in: body
        name: registerReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.RegisterReq'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功，返回用户信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或注册失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 用户注册
      tags:
      - 用户管理
...
  /api/v1/token:
    post:
      consumes:
      - application/json
      description: 验证用户提供的 token 是否有效
      parameters:
      - description: 验证 token 请求参数
        in: body
        name: verifyReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.VerifyTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: 验证成功，返回验证结果和 claims
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或验证失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 验证token
      tags:
      - 用户管理
...
  /api/v1/users/add:
    post:
      consumes:
      - application/json
      description: 管理员新增用户信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 新增用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysUser'
      produces:
      - application/json
      responses:
        "200":
          description: 新增成功，返回用户信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或新增用户失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 新增用户
      tags:
      - 用户管理
  /api/v1/users/del:
    post:
      consumes:
      - application/json
      description: 管理员删除用户信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除用户请求参数，包含用户 ID
        in: body
        name: deleteReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.DeleteByIdReq'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或删除用户失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 删除用户
      tags:
      - 用户管理
  /api/v1/users/get:
    get:
      consumes:
      - application/json
      description: 根据请求参数获取用户详细信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: loginName
        type: string
      - in: query
        name: mobile
        type: string
      - in: query
        name: nickName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或获取用户信息失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取用户信息
      tags:
      - 用户管理
  /api/v1/users/list:
    get:
      consumes:
      - application/json
      description: 根据请求参数获取用户分页列表
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - in: query
        name: deptId
        type: integer
      - in: query
        name: keyword
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: postId
        type: integer
      - in: query
        name: roleId
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: userType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回用户列表和分页信息
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.ResultPage'
        "500":
          description: 请求参数错误或获取用户列表失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 获取用户列表
      tags:
      - 用户管理
  /api/v1/users/logout:
    post:
      consumes:
      - application/json
      description: 用户退出当前登录状态
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 退出登录失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 退出登录
      tags:
      - 用户管理
  /api/v1/users/pwd:
    post:
      consumes:
      - application/json
      description: 用户修改自己的登录密码
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改密码请求参数
        in: body
        name: changePwdReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.ChangePasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或修改密码失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 修改密码
      tags:
      - 用户管理
  /api/v1/users/status:
    post:
      consumes:
      - application/json
      description: 管理员修改用户的状态
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改用户状态请求参数
        in: body
        name: changeStatusReq
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys_request.ChangeStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误或修改用户状态失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 修改状态
      tags:
      - 用户管理
  /api/v1/users/update:
    post:
      consumes:
      - application/json
      description: 管理员修改用户信息
      parameters:
      - description: 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 修改后的用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_maczh_mgin_models_sys.SysUser'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
        "500":
          description: 请求参数错误、用户 ID 为空或修改用户失败
          schema:
            $ref: '#/definitions/github_com_maczh_mgin_models.Result-any'
      summary: 修改用户
      tags:
      - 用户管理
swagger: "2.0"